# docs/swagger.yaml
basePath: /v1
definitions:
  appreviews.AppReview:
    properties:
      created_at:
        type: string
      feedback:
        type: string
      id:
        type: integer
      rating:
        type: integer
      user_id:
        type: integer
    type: object
...
  gameqa.Question:
    properties:
      created_at:
        type: string
      game_id:
        type: integer
      id:
        type: integer
      question:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  gameqa.QuestionWithReplies:
    properties:
      created_at:
        type: string
      id:
        type: integer
      question:
        type: string
      replies:
        items:
          $ref: '#/definitions/gameqa.Reply'
        type: array
    type: object
  gameqa.Reply:
    properties:
      admin_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      question_id:
        type: integer
      reply:
        type: string
      updated_at:
        type: string
    type: object
...
  main.AppReviewPayload:
    properties:
      feedback:
        type: string
      rating:
        type: integer
    type: object
...
  main.QuestionPayload:
    properties:
      question:
        maxLength: 120
        type: string
    required:
    - question
    type: object
...
  main.ReplyPayload:
    properties:
      reply:
        maxLength: 120
        type: string
    required:
    - reply
    type: object
...
  venues.Venue:
    properties:
      address:
        type: string
      amenities:
        description: Array of strings
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_urls:
        description: Array of image URLs
        items:
          type: string
        type: array
      location:
        description: PostGIS point (longitude, latitude)
        items:
          type: number
        type: array
      name:
        type: string
      open_time:
        type: string
      owner_id:
        type: integer
      phone_number:
        type: string
      sport:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: Ongchen10sherpa@gmail.com
    name: fullstacksherpa
    url: https://www.fullstacksherpa.tech/
  description: API for Khel, a complete sport application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Khel API
paths:
  /app-reviews:
    get:
      consumes:
      - application/json
      description: Returns all app reviews from all users.
      produces:
      - application/json
      responses:
        "200":
          description: List of all reviews
          schema:
            items:
              $ref: '#/definitions/appreviews.AppReview'
            type: array
        "401":
          description: unauthorized route
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: List All App Reviews
      tags:
      - App_Reviews
    post:
      consumes:
      - application/json
      description: Allows authenticated users to submit a rating and feedback for
        the app.
      parameters:
      - description: Review payload
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/main.AppReviewPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Review submitted
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Submit App Review
      tags:
      - App_Reviews
...
  /games/{gameID}/qa:
    get:
      consumes:
      - application/json
      description: Get all questions with replies for a game
      parameters:
      - description: Game ID
        in: path
        name: gameID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gameqa.QuestionWithReplies'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get all Q&A for a game
      tags:
      - Questions
  /games/{gameID}/questions:
    get:
      consumes:
      - application/json
      description: Get all questions for a game
      parameters:
      - description: Game ID
        in: path
        name: gameID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gameqa.Question'
            type: array
        "400":
          description: Bad Request
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get game questions
      tags:
      - Questions
    post:
      consumes:
      - application/json
      description: Create a new question for a game
      parameters:
      - description: Game ID
        in: path
        name: gameID
        required: true
        type: integer
      - description: Question payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.QuestionPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gameqa.Question'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a game question
      tags:
      - Questions
  /games/{gameID}/questions/{questionID}:
    delete:
      consumes:
      - application/json
      description: Delete a question (only by author)
      parameters:
      - description: Game ID
        in: path
        name: gameID
        required: true
        type: integer
      - description: Question ID
        in: path
        name: questionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete a question
      tags:
      - Questions
  /games/{gameID}/questions/{questionID}/replies:
    post:
      consumes:
      - application/json
      description: Create a reply to a question (admin only)
      parameters:
      - description: Game ID
        in: path
        name: gameID
        required: true
        type: integer
      - description: Question ID
        in: path
        name: questionID
        required: true
        type: integer
      - description: Reply payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.ReplyPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gameqa.Reply'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a reply
      tags:
      - Questions
...
  /venues/{venueID}/favorite:
    delete:
      consumes:
      - application/json
      description: Allows authenticated users to remove a venue from their favorites
        list.
      parameters:
      - description: Venue ID
        in: path
        name: venueID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Venue removed from favorites
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Invalid venue ID or unauthenticated request'
          schema: {}
        "500":
          description: 'Internal Server Error: Could not remove favorite'
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Remove a venue from favorites
      tags:
      - Favorite_Venues
    post:
      consumes:
      - application/json
      description: Allows authenticated users to add a venue to their favorites list.
      parameters:
      - description: Venue ID
        in: path
        name: venueID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Venue added to favorites
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Invalid venue ID or unauthenticated request'
          schema: {}
        "500":
          description: 'Internal Server Error: Could not add favorite'
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Add a venue to favorites
      tags:
      - Favorite_Venues
...
  /venues/favorites:
    get:
      consumes:
      - application/json
      description: Returns a list of venues that the authenticated user has marked
        as favorites.
      produces:
      - application/json
      responses:
        "200":
          description: List of favorite venues
          schema:
            items:
              $ref: '#/definitions/venues.Venue'
            type: array
        "400":
          description: 'Bad Request: Unauthenticated request'
          schema: {}
        "500":
          description: 'Internal Server Error: Could not retrieve favorites'
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Retrieve user's favorite venues
      tags:
      - Favorite_Venues
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"